Lukkien Test/requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
djangorestframework==3.9.1
python-decouple==3.1
pytz==2018.9
static3==0.7.0
gunicorn
psycopg2
